+++ src/DesktopGrass.Native/src/App.h
@@ -15,6 +15,7 @@
 #include "GrassWindow.h"
 #include "MouseHook.h"
 #include "Persistence.h"
+#include "Config.h"
 
 namespace desktopgrass {
 
@@ -83,6 +84,7 @@ private:
     MouseEventQueue                             queue_{};
     std::vector<std::unique_ptr<GrassWindow>>   windows_;
     uint64_t                                    seed_     = 0;
+    config::Config                              config_{};
     Scene                                       currentScene_ = SCENE_DEFAULT;
     CritterKind                                 currentCritter_ = CRITTER_DEFAULT;
     int                                         currentCritterCount_ = 0;